TT#105151 Stop adding "net.ifnames=0" to grub config
This is commit cd50e4934cf4ab80149420d66cbce5191f620d07 applied again.

As explained in ab62171c49c46fae2b7150ebc9e9d2e0a36843e1, the original
change had to be reverted because even if things work perfectly fine, in
the case of Vagrant machines (or when passing "vagrant" parameter to the
script) the udev-rules for persistent-net devices get removed, so then
the network interfaces get "random" names and the configuration in
/etc/network/interfaces doesn't match, the network is not brought up.

This removal happens in the case of {ce,pro,carrier}-trunk.mgm machines
of our tests, which shouldn't be needed, and also in the images created
for Vagrant machines, which is understandable because the machines could
be brought up with different PCIIDs in different versions of VirtualBox,
or due to some other difference -- not sure how we can ensure that the
PCIIDs as written in the udev-rules files will work in that case.

But in principle this change must go ahead when we solve these problems,
so submitting it again to be ready.

Change-Id: Ib39481a2608aa56e6ec6c9255e290787a6ce3af7
(cherry picked from commit a50903a30c5710529046bc19196ed42ce07e78b0)