TT#105151 Do the renaming eth*->neth* outside of the "if $NGCP_INSTALLER" block
Jobs like daily-build-matrix-debian-boxes build plain Debian machines,
not NGCP-based ones.  At the moment we're generating the udev-rules for
network renaming unconditionally, so we have to do it consistently,
either both conditionally and not for "plain" systems, or both
unconditionally, so network can be brought up by a correct
/etc/network/interfaces after the devices are brought up with the new
names.

There is a good-ish argument for keeping using eth0, as it is more of a
default, but we're already deviating from the default for several years
and Debian stable releases by having these names and not ones like
"ens18" or "enp4s0f2" which is the default in Debian nowadays, at least
since buster.

So it is probably better to keep it consistent with our other machines
and use "neth*" naming for those too.

Change-Id: I6b3b49a1769894580df768abb817ae5196e65963