MT#57559 Always include zstd when bootstrapping systems
During initial deployment of a system, we get warnings about
lack of zstd:

| Setting up linux-image-6.1.0-13-amd64 (6.1.55-1) ...
| I: /vmlinuz.old is now a symlink to boot/vmlinuz-6.1.0-13-amd64
| I: /initrd.img.old is now a symlink to boot/initrd.img-6.1.0-13-amd64
| I: /vmlinuz is now a symlink to boot/vmlinuz-6.1.0-13-amd64
| I: /initrd.img is now a symlink to boot/initrd.img-6.1.0-13-amd64
| /etc/kernel/postinst.d/initramfs-tools:
| update-initramfs: Generating /boot/initrd.img-6.1.0-13-amd64
| W: No zstd in /usr/bin:/sbin:/bin, using gzip
| [...]

The initramfs generation and update overall runs *four* times within the
initial bootstrapping of a system (we'll try to do something about this,
but this is outside the scope of this).

As of initramfs-tools v0.141, initramfs-tools uses zstd as default
compression for initramfs. Version 0.142 is shipped with
Debian/bookworm, and therefore it makes sense to have it available
upfront. Note that also the initrd generation is faster with zstd
(~10sec for zstd vs. ~13sec for gzip) and also the resulting initrd is
smaller (~33MB for zstd vs ~39MB for gzip).

By making sure that zstd is available straight from the very beginning
and before ngcp-installer pulls it in later, we can avoid the warning
message but also save >10 seconds of install time.

Given that zstd is available even in Debian oldoldstable, let's install
it unconditionally in all our systems.

Thanks: Volodymyr Fedorov for reporting

Change-Id: I56674c3c213f7c7a6e6cbce3c8e2e00a4cfbdbd4