MT#58014 Remove support for fetching OpenPGP certificates from keyservers
The code trying to fetch the OpenPGP certificate from a keyserver has
been non-functional for a while as the GPG_KEY_SERVER variable was
removed in commit 316c28bcc2c0a4eb3d7b5f457cb4cd8c30f54695. Instead of
restoring the variable with an up-to-date keyserver (not part of the
SKS pool, as that network is dead), we remove the support entirely as
it's a potential security issue due to fingerprint collisions for
example.

As a side effect this removes apt-key usage which has been deprecated
upstream and is slated for removal.

Change-Id: I63171a66201c631da9233d54579bd1601ff22e3e