MT#59872 Remove NGCP_PXE_INSTALL variable
With this variable we had some tricks in ngcp-initial-configuration if
the Pro sp2 node is installer via iPXE/cm image.

Now we support installation of sp2 via iPXE only so no need to pass this
variable.

But we need to keep parent ngcppxeinstall parameter as we need this
information for netcardconfig.

Change-Id: I20491289917cbb427ad6f5670f108c632838be71