MT#62787 Drop /sys/block autodetection code
This was actually hiding a problem and is a f'up related to the previous
commit 001c9c607331bf501e208b5a7ed9625502fc24dd.

When we're deploying a system with SW-RAID to sd* based disks (with boot
options like swraiddisk1=sda swraiddisk2=sdb), we're not setting
TARGET_DISK, but instead old autodetection code kicks in.  This only
supports block devices like vd* + sd*, but then sets DISK to the first
such present disk.

As a result, we're ending up with a grml-debootstrap command line like
this:

| (netscript.grml:2219):  grml-debootstrap --arch amd64 --grub /dev/sda --filesystem ext4 [...] -r bookworm -t /dev/mapper/ngcp-root --efi /dev/sda2 [...]

This propagates through grml-debootstrap into GRUB's debconf settings
for the /dev/sda disk, showing up like this:

| root@db01a:~# debconf-show grub-cloud-amd64
| * grub-cloud-amd64/install_devices: /dev/disk/by-id/ata-MTFDDAK480TDT-1AW1ZA_02JG544D7A44709LEN_3416E3FC

This is unexpected behavior, as GRUB would get installed on one disk only.

Also, when deploying to disks other than vda and sd* (e.g. with
NVMe ones), this shouldn't just use unrelated disks, but instead either
work with the expected ones or abort hard (as has been observed in the
previous commit with the NVMe disks).

Furthermore there's no point in further shipping this autodetection code
any longer at all, as we're using the TARGET_DISK approach only via
disk_selection.sh. Otherwise we're expecting the deployment system
(Grml-Deployment ISO, PXE boot environment,...) to fully control and
pass expected disk options.

It's 2025, and NVMe disks are a thing, so good riddance!

Thanks: Volodymyr Fedorov for reporting and assistance

Change-Id: I3fd7aa75a38dc0f07bc07d5b22139a89498f7b96